requirements.txt
telethon==1.23.0
python-dotenv==0.17.0